# Use a hash table to index existing vtables