This allows for quick deduplication even in situations where there might be thousands of vtables due to 'combinatoric explosion'.

This changeset targets Python only.

This fixes issue #5301.