@aardappel 2019-05-02
Thanks!

Not being a Python programmer, my C++ intuition tells me that vtable dedup should be linear until N (where N is determined by experimentation) and a map only after, since there is overhead associated with a map, particularly in memory allocation. The C++ implementation painstakingly makes the vtable list part of the same one buffer everything else sits in.

But that is C++. It may well be that in Python going with a map from the start is faster, since it avoids a lot of the handwritten loops we had before, which are bound to be slow in Python. But I'd like some numbers. Can you write a quick benchmark (not part of this PR) where using this PR and the old code, you create between 1 and N unique vtables of some reasonable length (say 5 entries), then look up those same vtables a gazillion times? Then find the break-even point?

Who knows the map is always faster, which would be great. Maybe the break-even point is like.. 3, in which case map-only is still acceptable. But if the break-even point is like, 10, we may want to map to only kick in after we've seen 10+ vtables (which is what I would expect in C++, but likely not Python).

@rw can you please review for Python correctness?