@malthe 2019-05-02
@aardappel – I'll try to find time to write up a quick benchmark, but in my own test case (which is millions of objects), there was absolutely no difference for N <= 5. And the benchmark that already comes with the implementation also showed a minor improvement in favor of the hash table.

I think there's a significant overhead in the vtable equality test that really cancels out the cost of the hash table. In C++ this test is presumably a lot more efficient.

Note that I have not really thought much about optimization since the original implementation already leaves quite a lot on the table.