@rw 2019-05-03
@malthe This is cool, thanks for making the PR!

Could you describe how memory usage will be affected? Because you're storing the entire vtable layout as a key in the hashmap, it seems to me that memory usage could get quite large, if there are many tables with varying structure. (IIRC a vtable can have the same keys, yet different byte layouts, due to alignment issues.)

WDYT about combining your approach with the existing approach, and use a hashmap to store offsets to vtables that we know have unique structure?