@rw 2019-05-03
@malthe Also like @aardappel I'd really like to see benchmarks!