@malthe 2019-05-03
I did try first using basically `hash(vtable)` (an integer) as the key for the map and then having the vtable candidates as the value, but Python's builtin hash-function for tuples of integers turned out to result in quite a few collisions.

A better hash function would improve on that, but then you have the overhead of having to do the equality test for the candidates (even if there's just one candidate.)

In my test case, I have some 13,000 vtables. I think the memory overhead is acceptable, because I'm also adding dozens of millions of objects. The size of the vtables themselves is still manageable.