@malthe 2019-05-03
This is the result of the updated benchmark.

Before:
```
vtable deduplication rate (n=1, vtables=1): 90788.67 sec
vtable deduplication rate (n=10, vtables=14): 17287.12 sec
vtable deduplication rate (n=100, vtables=125): 4193.43 sec
vtable deduplication rate (n=1000, vtables=1031): 644.62 sec
```

After:
```
vtable deduplication rate (n=1, vtables=1): 118095.27 sec
vtable deduplication rate (n=10, vtables=11): 50684.67 sec
vtable deduplication rate (n=100, vtables=129): 25768.73 sec
vtable deduplication rate (n=1000, vtables=1038): 17963.06 sec
```
That is, the move to a hash table always makes writing the vtable faster, but it very quickly becomes a lot faster as the number of vtables increase.