@aardappel 2019-05-06
To be clear on what your numbers mean, the final number is "iterations per second" ? And what is `n`?

Agree that rather than storing the whole vtable, it would be better to store the index of the vtable, but that would require a custom hashing/comparing function.. which Python does not support? Or it would be slow?