@rw 2019-05-06
IIRC `n` means how many fields there are in the table, and `vtables` is how many unique vtables there are in the buffer.

One benefit of this proposed approach is that the algorithmic complexity is definitely better: `O(1)` instead of `O(num_vtables)`.

One problem of this proposed approach is that if the dictionary tuples are large, or if the dictionary has many elements, then the memory usage will be closer to `O(num_vtables)` instead of `O(1)`.

Overall I'm in favor of this change, because our users are likely to be using Python in contexts where they would prefer to use more memory to get much faster speeds.

In the future, we could customize the hash table implementation to make it better for our particular use case. This could include making the vtable deduplicator be approximate instead of exact, creating a custom hash function, and/or storing only the offsets into the buffer instead of the vtable tuple.