@aardappel 2019-05-06
Ok, then I don't quite understand why `n` has to vary so much, but it looks unlikely that different benchmarking would invalidate the result.

Since @rw is cool with this, lets merge this, and see if we can improve memory usage down the line. I think large numbers of vtables is rare though, so hopefully it doesn't matter.