@rw 2019-05-06
Concretely, I would implement improve this by doing the following:

1) Create a hashmap of the type: `int -> List[Offset]`.
2) Implement a simple hash function that turns a vtable tuple into an integer.
3) Use that unsigned integer as a key into the hashmap, to fetch the list of offsets that hash to that key.
4) Use the list of offsets to find candidate vtables to use for deduplication, as in the original algorithm before this PR.

As a result, the memory usage will only involve storing a hashmap of ints and lists of ints, not the full vtable tuples.