@malthe 2019-05-06
Actually, `n` is not the number of fields in the table – it's the number of logical vtables in the test case. It used to be the number of fields in the table, but I changed it to make it more realistic. That is, in order to show the effect I was seeing in real life, I had to demonstrate the "combinatoric explosion".