@rw 2019-05-06
@malthe Would you make a followup PR adding some more comments to the code? I saw you deleted the block comment I wrote explaining the algorithm, but did not replace it with a new block comment.