@malthe 2019-05-06
@rw, I have some data points for that approach.

It's possible to reduce memory usage by "packing" the vtable. In the common case, this will significantly reduce the memory overhead, typically 2-3 times less. The downside is that it adds to the code complexity.
```diff
--- a/python/flatbuffers/builder.py
+++ b/python/flatbuffers/builder.py
@@ -192,9 +192,11 @@ class Builder(object):
 
         objectOffset = self.Offset()
 
-        vtKey = []
         trim = True
-        for elem in reversed(self.current_vtable):
+        bw = N.VOffsetTFlags.bytewidth * 8
+        bs = (len(self.current_vtable) - VtableMetadataFields) * bw
+        i = 0
+        for j, elem in enumerate(reversed(self.current_vtable)):
             if elem == 0:
                 if trim:
                     continue
@@ -202,9 +204,9 @@ class Builder(object):
                 elem = objectOffset - elem
                 trim = False
 
-            vtKey.append(elem)
+            i += bs ** j * elem
 
-        vtKey = tuple(vtKey)
+        vtKey = 2 ** bw * i + bs
         vt2Offset = self.vtables.get(vtKey)
```
To reduce memory usage further, one might instead use `hashlib.md5` as the hash function, shortening to 64 bits and then run the required equality test. This adds a bit of computation overhead.