@rw 2019-05-06
@malthe I'd build the vtkey as we already do, but just hash that into an integer instead of using it as the hashmap key. The hash function can be pretty simple -- if it's bad, we'll just have more collisions, and if it's good, then everything will be fast.